requirements.txt
flask
pandas
scikit-learn
imbalanced-learn
xgboost
joblib
gspread
google-auth
python-dotenv
